financial에서 카드 버전, 실적, 혜택 데이터 가져오는 api 추가

+++ Backend/financial/src/main/java/com/financial/domain/bank/usercard/model/dto/response/BankUserCardGetResponse.java
@@ -17,7 +17,7 @@ import lombok.ToString;
 public class BankUserCardGetResponse {
 
     @NotNull
-    private Long bankUserCardId;
+    private int bankUserCardId;
 
     @NotBlank
     private String bankCardId;
